Moved Amd1ACpu Zen 5 handling into Amd17Cpu, Fixed typo
