+++ LibreHardwareMonitorLib/Hardware/RyzenSMU.cs
@@ -159,17 +159,17 @@ internal class RyzenSMU
                 { 11, new SmuSensorType { Name = "Package", Type = SensorType.Temperature, Scale = 1 } },
                 { 595, new SmuSensorType { Name = "L3 (CCD1)", Type = SensorType.Temperature, Scale = 1 } },
                 { 596, new SmuSensorType { Name = "L3 (CCD2)", Type = SensorType.Temperature, Scale = 1 } },
-                { 52, new SmuSensorType { Name = "VDDCR_VDD", Type = SensorType.Temperature, Scale = 1 } },
-                { 57, new SmuSensorType { Name = "VDDCR_SOC", Type = SensorType.Temperature, Scale = 1 } },
-                { 62, new SmuSensorType { Name = "VDD_MISC", Type = SensorType.Temperature, Scale = 1 } },
+                { 52, new SmuSensorType { Name = "VDDCR VDD", Type = SensorType.Temperature, Scale = 1 } },
+                { 57, new SmuSensorType { Name = "VDDCR SoC", Type = SensorType.Temperature, Scale = 1 } },
+                { 62, new SmuSensorType { Name = "VDD Misc", Type = SensorType.Temperature, Scale = 1 } },
                 { 21, new SmuSensorType { Name = "SoC Power", Type = SensorType.Power, Scale = 1 } },
                 { 54, new SmuSensorType { Name = "SoC Voltage", Type = SensorType.Voltage, Scale = 1 } },
                 { 55, new SmuSensorType { Name = "SoC Current", Type = SensorType.Current, Scale = 1 } },
                 { 71, new SmuSensorType { Name = "Fabric", Type = SensorType.Clock, Scale = 1 } },
                 { 75, new SmuSensorType { Name = "Uncore", Type = SensorType.Clock, Scale = 1 } },
                 { 79, new SmuSensorType { Name = "Memory", Type = SensorType.Clock, Scale = 1 } },
-                { 58, new SmuSensorType { Name = "VDD_Misc", Type = SensorType.Voltage, Scale = 1 } },
-                { 83, new SmuSensorType { Name = "VDDCR_SoC", Type = SensorType.Voltage, Scale = 1 } },
+                { 58, new SmuSensorType { Name = "VDD Misc", Type = SensorType.Voltage, Scale = 1 } },
+                { 83, new SmuSensorType { Name = "VDDCR SoC", Type = SensorType.Voltage, Scale = 1 } },
                 { 259, new SmuSensorType { Name = "CLDO VDDG IOD", Type = SensorType.Voltage, Scale = 1 } },
                 { 261, new SmuSensorType { Name = "CLDO VDDG CCD", Type = SensorType.Voltage, Scale = 1 } },
                 { 269, new SmuSensorType { Name = "CLDO VDDP", Type = SensorType.Voltage, Scale = 1 } },
